修复: 注入顺序错误导致应用启动失败.
