+++ wind-core/src/main/java/io/github/ramerf/wind/core/config/CommonBean.java
@@ -1,14 +1,90 @@
 package io.github.ramerf.wind.core.config;
 
+import io.github.ramerf.wind.core.converter.TypeConverter;
+import io.github.ramerf.wind.core.entity.enums.InterEnum;
+import io.github.ramerf.wind.core.executor.Executor;
+import io.github.ramerf.wind.core.executor.JdbcTemplateExecutor;
+import io.github.ramerf.wind.core.factory.TypeConverterRegistryFactory;
+import io.github.ramerf.wind.core.serializer.JacksonEnumSerializer;
+import io.github.ramerf.wind.core.support.StringToEnumConverterFactory;
+import java.util.LinkedHashSet;
+import java.util.Set;
+import javax.annotation.Nonnull;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 /**
+ * 定义常用bean.
+ *
  * @author Tang Xiaofeng
  * @since 2019/12/29
  */
 @Slf4j
 @Configuration("wind_core_common_bean")
 public class CommonBean {
-  // 定义常用bean
+  @Autowired(required = false)
+  @SuppressWarnings({"rawtypes", "SpringJavaAutowiredFieldsWarningInspection"})
+  private final Set<TypeConverter> typeConverters = new LinkedHashSet<>();
+
+  /**
+   * Type converter registry factory type converter registry factory.
+   *
+   * @return the type converter registry factory
+   */
+  @Bean
+  public TypeConverterRegistryFactory typeConverterRegistryFactory() {
+    final TypeConverterRegistryFactory factory = new TypeConverterRegistryFactory();
+    factory.addTypeConverter(typeConverters);
+    factory.registerDefaultTypeConverters();
+    return factory;
+  }
+
+  /**
+   * String to enum converter factory mvc configure web mvc configurer.
+   *
+   * @return the web mvc configurer
+   */
+  @Bean
+  public WebMvcConfigurer stringToEnumConverterFactoryMvcConfigure() {
+    // 添加枚举转换器,请求可以传递value整型值
+    return new WebMvcConfigurer() {
+      @Override
+      public void addFormatters(@Nonnull FormatterRegistry registry) {
+        registry.addConverterFactory(new StringToEnumConverterFactory());
+      }
+
+      @Override
+      public void addCorsMappings(@Nonnull CorsRegistry registry) {
+        final long maxAge = 3600L;
+        registry
+            .addMapping("/**")
+            .allowedOrigins("*")
+            .allowedMethods("GET", "HEAD", "POST", "PUT", "PATCH", "DELETE", "OPTIONS", "TRACE")
+            .allowCredentials(false)
+            .maxAge(maxAge);
+      }
+    };
+  }
+
+  /**
+   * Jackson object mapper customizer jackson 2 object mapper builder customizer.
+   *
+   * @return the jackson 2 object mapper builder customizer
+   */
+  @Bean
+  public Jackson2ObjectMapperBuilderCustomizer jacksonObjectMapperCustomizer() {
+    return objectMapperBuilder ->
+        objectMapperBuilder.serializerByType(InterEnum.class, new JacksonEnumSerializer());
+  }
+
+  @Bean
+  public Executor jdbcTemplateExecutor() {
+    return new JdbcTemplateExecutor();
+  }
 }
